requirements.txt
streamlit==1.27.0
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.1.154
chromadb==0.3.30
langchain-google-genai==0.0.14
python-dotenv==1.0.1
